+++ packages/config/src/xcm-configs/alphanetRelay.ts
@@ -28,7 +28,7 @@ export const alphanetRelayRoutes = new ChainRoutes({
       },
       extrinsic: ExtrinsicBuilder()
         .xcmPallet()
-        .limitedReserveTransferAssets(0)
+        .transferAssetsUsingTypeAndThen()
         .here(),
     },
   ],